Refactor logic into separate classes
